fix: Update import paths after analysis-plugins directory move
- Fix import paths in commentGenerator.ts, extractionHelper.ts, and pluginHelpers.ts
- Remove unused documentChunking.ts file
- Resolves TypeScript import errors after plugin directory reorganization
